requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
openai>=1.30
python-dotenv>=1.0.0
pillow>=10.0.0
